fix: DTO's not being read for prescrption
The issue was due to no presence of setter and getter method for the attributes of the prescription which was preventing the system from picking up the fields sent by from the url.
The names of the attributes are also changed to Pascal case as per the convetion
Some of the attributes of the prescription are set to required.
Removed not neccessary folders to prevent them from pushing accidenlty